compiled map onto base.html and shortened file tree, set up basic routes for member, still needs work.
