--- app/templates/index.html
@@ -1,19 +1,0 @@
-{% extends "base.html" %}
-{% block title %}Home{% endblock %}
-{% block body %}
-<div class="jumbotron">
-  <h1>Hello, Dear Visitor!</h1>
-</div>
-  <p class="lead">Welcome to the NYU Diploma in Python Programming course website. We are a team of amazing individuals with different backgrounds who love to code and love Python. We started this course in June 2017 and already know so much. We built this site ourselves!  </p>
-  <p class="lead">The course consists of 3 classes about programming in Python and a capstone project which we will devote the last class to. In the capstone class, each of us will build an independent project in Python and make it available on our public GitHub repositories. Feel free to checkout! </p>
-  <p class="lead">You can also checkout our profiles and completed projects in Members and Projects links in the navbar, respectively. </p>
-<img src="{{url_for('static', filename=['MAPFILE')}}" />
-{% endblock %}
-
-
-
-
-
-
-
-
