--- app/templates/map3.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<html>
-<head>
-<link rel="stylesheet" type="text/css" href="map3.css">
-<script src="https://code.jquery.com/jquery-1.10.2.js"></script>
-<script src="map3.js"></script>
-</head>
-
-<body>
-  <button type="submit" style="background: url(test.jpg);"></button>
-  <button type="submit" style="background: url(test.jpg);"></button>
-  <button type="submit" style="background: url(test.jpg);"></button>
-  <button type="submit" style="background: url(test.jpg);"></button>
-  <button type="submit" style="background: url(test.jpg);"></button>
-  <button type="submit" style="background: url(test.jpg);"></button>
-  <button type="submit" style="background: url(test.jpg);"></button>
-  <button type="submit" style="background: url(test.jpg);"></button>
-</body>
-
-</html>
