+++ app/templates/member.html
@@ -1,0 +1,1 @@
+<!--This page will have templating for each class member-->
