+++ app/templates/projects.html
@@ -1,0 +1,1 @@
+<!--This page will have templating for each project-->
