--- content/sarah_black.md
@@ -1,31 +1,0 @@
-# Sarah Black
-
-## Bio
-Project manager; (crappy but overjoyed) dancer; (lazy) mma fighter; feels like she's writing a dating profile
-
-## Motivation
-My boss told me to, but also I wanted to. Got a lot of reporting that needs to be automated and no one else is going to do it for me.
-
-## Area of interest
-Not sure at the moment - I'll see what I can do with Python at my job and just sort of see where it takes me. I'm not looking to go into a new field, just looking to have more tools at my disposal to do my current job better.
-
-## Fun Facts
-1. I grew up in Rye, 45 minutes outside the city
-2. I love to dance but in the ~26 years I've been doing it, I've never been very good
-3. I would subsist on carbs, cheese, and burgers alone if it were possible
-4. Gif not jif
-5. I genuinely believe that seltzer is better than regular water
-
-## Favorite movie
-I hate this question. I don't have one favorite movie, it completely depends on my mood. But some that are in my list of favorites that I've recently watched are: Her, Bone Collector, The Avengers, Charlie's Angels, again, I have very wide tastes.
-
-## Dream project
-I don't have dreams.
-
-Aka I just want to figure out the stuff I know I want to do, haven't thought much beyond that.
-
-## Favorite Disney Princess
-[#arielfolyfe](https://www.youtube.com/watch?v=SXKlJuO07eM "Part of Your World")
-
-## Github
-[sarah-loring](https://github.com/sarah-loring/nyu-python "github")
