--- navbar/templates/base.html
@@ -1,52 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>{% block title %}{% endblock %}</title>
-    <!-- Bootstrap core CSS -->
-    <link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" rel="stylesheet">
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-    <style type="text/css">
-      .my-nav{
-        margin: 20px;
-      }
-    </style>
-    <!-- Custom styles for this template -->
-    <link href="https://getbootstrap.com/examples/jumbotron-narrow/jumbotron-narrow.css" rel="stylesheet">
-  </head>
-  <body>
-    <div class="container">
-      <div class="my-nav">
-        <nav>
-          <ul class="nav nav-pills">
-            <li role="presentation"><img src="/static/icon.png"></li>
-            <li class="dropdown">
-              <a href="#" data-toggle="dropdown" class="dropdown-toggle">Members <b class="caret"></b></a>
-              <ul class="dropdown-menu">
-                  <li><a href="/alice">Alice</a></li>
-                  <li><a href="/ana">Ana</a></li>
-                  <li><a href="/dean">Dean</a></li>
-                  <li><a href="/isobel">Isobel</a></li>
-                  <li><a href="/kai">Kai</a></li>
-                  <li><a href="/misha">Misha</a></li>
-                  <li><a href="/sarah">Sarah</a></li>
-              </ul>
-            </li>
-            <li class="dropdown">
-              <a href="#" data-toggle="dropdown" class="dropdown-toggle">Instructor <b class="caret"></b></a>
-              <ul class="dropdown-menu">
-                  <li><a href="/jessica">Jessica</a></li>
-              </ul>
-            </li>
-            <li role="presentation"><a href="/projects">Projects</a></li>
-          </ul>
-        </nav>
-      </div>
-      {% block content %}
-      {% endblock %}
-      <footer class="footer">
-        <p>NYU Advanced Python Class</p>
-      </footer>
-    </div> <!-- /container -->
-  </body>
-</html>
